refactor: ReservationTime 메서드 위치 변경
