+++ src/main/java/roomescape/reservation/controller/ReservationTimeController.java
@@ -1,10 +1,13 @@
 package roomescape.reservation.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.reservation.dto.AvailableReservationTimeResponse;
 import roomescape.reservation.dto.ReservationTimeResponse;
 import roomescape.reservation.service.ReservationTimeService;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @RestController
@@ -21,6 +24,19 @@ public class ReservationTimeController {
         return reservationTimeService.getReservationTimes()
                 .stream()
                 .map(ReservationTimeResponse::from)
+                .toList();
+    }
+
+    @GetMapping("/available-reservation-times")
+    public List<AvailableReservationTimeResponse> getAvailableReservationTimes(
+            @RequestParam("date") final LocalDate date,
+            @RequestParam("theme-id") final Long themeId
+    ) {
+        return reservationTimeService.getAvailableReservationTimes(date, themeId)
+                .values()
+                .entrySet()
+                .stream()
+                .map(entry -> AvailableReservationTimeResponse.of(entry.getKey(), entry.getValue()))
                 .toList();
     }
 }
