+++ src/main/java/roomescape/reservation/controller/ThemeController.java
@@ -1,13 +1,10 @@
 package roomescape.reservation.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.reservation.dto.AvailableReservationTimeResponse;
 import roomescape.reservation.dto.ThemeResponse;
 import roomescape.reservation.service.ThemeService;
 
-import java.time.LocalDate;
 import java.util.List;
 
 @RestController
@@ -32,16 +29,6 @@ public class ThemeController {
         return themeService.getPopularThemes()
                 .stream()
                 .map(ThemeResponse::from)
-                .toList();
-    }
-
-    @GetMapping("/available-reservation-times")
-    public List<AvailableReservationTimeResponse> getAvailableReservationTimes(@RequestParam("date") final LocalDate date, @RequestParam("theme-id") final Long themeId) {
-        return themeService.getAvailableReservationTimes(date, themeId)
-                .values()
-                .entrySet()
-                .stream()
-                .map(entry -> AvailableReservationTimeResponse.of(entry.getKey(), entry.getValue()))
                 .toList();
     }
 }
